:truck: Moved network state logic to core
