+++ base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/activity/climb/AreaActivity.kt
@@ -13,6 +13,7 @@ import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.ensureGet
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.area.has
 import com.arnyminerz.escalaralcoiaicomtat.core.data.climb.zone.Zone
 import com.arnyminerz.escalaralcoiaicomtat.core.exception.NoInternetAccessException
+import com.arnyminerz.escalaralcoiaicomtat.core.network.base.ConnectivityProvider
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.AREAS
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_AREA
 import com.arnyminerz.escalaralcoiaicomtat.core.shared.EXTRA_AREA_TRANSITION_NAME
@@ -26,7 +27,6 @@ import com.arnyminerz.escalaralcoiaicomtat.core.utils.put
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.putExtra
 import com.arnyminerz.escalaralcoiaicomtat.core.utils.uiContext
 import com.arnyminerz.escalaralcoiaicomtat.list.model.dwdataclass.DwDataClassAdapter
-import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
 import timber.log.Timber
 
 /**
