--- base/src/main/java/com/arnyminerz/escalaralcoiaicomtat/shared/SharedStates.kt
@@ -1,8 +1,0 @@
-package com.arnyminerz.escalaralcoiaicomtat.shared
-
-import com.arnyminerz.escalaralcoiaicomtat.network.base.ConnectivityProvider
-
-var appNetworkState: ConnectivityProvider.NetworkState =
-    ConnectivityProvider.NetworkState.NOT_CONNECTED
-
-lateinit var appNetworkProvider: ConnectivityProvider
